extract MatrixNode and MatrixLightClientNode classes from transport.py into node.py
